1.5
Reworked the way the card element grid uses the filters to determine which cards should show up.  Uses a refreshGrid function every time a button is clicked that evaluates the filters in a more efficient way, looking at cards one at a time to decide if they should be displayed or not.