3.3
Made a bunch of changes. 
Reworked the html from the ground up and then rebuild the CSS and JS to fit it.
Has most of the same functions as before.
Added a 'flip' button, rather than making the whole card perform the flip action.
Changed the coloring to be more sleek and consistent.